HEXBOX-284
The filtering issue on “All” tab was solved.

There was an issue with the search input: if the query was manually deleted instead of using the “X” button, the tabs couldn't filter the campaigns properly afterward, as the query still persisted in the URL. This issue has now been resolved.